+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- `textDecorationLine` in TextWidget style
+
 ## [0.14.0] - 2024-05-13
 
 ### Added